Create option to change pixel density of canvas
Had I used prompt, the user would have to type something whenever the
canvas is cleared, which can be annoying, so I'm reading it from the
input field.

The need to completely redraw the canvas due to density change
deprecated the function "clear canvas" (no need to clear something that
will be completely wiped out later).